[Todo API][Bryan-Vance]feature: add exception to handle update when given wrong id

+++ src/main/java/com/oocl/todolistapi/service/TodoService.java
@@ -1,10 +1,13 @@
 package com.oocl.todolistapi.service;
 
+import com.oocl.todolistapi.exception.TodoNotFoundException;
 import com.oocl.todolistapi.model.Todo;
 import com.oocl.todolistapi.repository.TodoRepository;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
+
+import static java.text.MessageFormat.format;
 
 @Service
 public class TodoService {
@@ -24,15 +27,19 @@ public class TodoService {
 
     public Todo update(Integer id, Todo updatedTodo) {
         Todo todo = todoRepository.findById(id).orElse(null);
-        if (todo != null){
+        if (todo != null) {
             updatedTodo.setId(id);
             updatedTodo.setText(todo.getText());
             return todoRepository.save(updatedTodo);
         }
-        return null;
+        throw new TodoNotFoundException(format("{0} not found", id));
     }
 
     public void deleteById(Integer id) {
+        Todo todo = todoRepository.findById(id).orElse(null);
+        if (todo == null) {
+            throw new TodoNotFoundException(format("{0} not found", id));
+        }
         todoRepository.deleteById(id);
     }
 }
